+++ WCIIRemake/cordScr.cpp
@@ -56,7 +56,7 @@ float cordScr::lineLength(cordScr cord1, cordScr cord2) {
 	if (cord1.x < 0 || cord1.y < 0 || cord2.x < 0 || cord2.y < 0) {
 		return -1;
 	}
-	return sqrt(sqr(cord2.x - cord1.x) + sqr(cord2.y - cord1.y));
+	return (float)sqrt(sqr(cord2.x - cord1.x) + sqr(cord2.y - cord1.y));
 }
 
 int cordScr::sqr(int input) {
